refactor: rename LL1 and sub units for consistency
