--- src/plcc/load_spec/validate_spec/validate_syntactic_spec/LL1/__init__.py
@@ -1,1 +1,0 @@
-from .LL1Validator import check_ll1
