+++ src/plcc/load_spec/validate_spec/validate_syntactic_spec/validate_ll1/__init__.py
@@ -1,0 +1,1 @@
+from .validate_ll1 import validate_ll1
