+++ src/plcc/load_spec/validate_spec/validate_syntactic_spec/validate_ll1/create_spec_grammar_test.py
@@ -2,7 +2,7 @@ from pytest import raises
 from plcc.load_spec.load_rough_spec.parse_dividers import parse_dividers
 from plcc.load_spec.load_rough_spec.parse_lines import Line
 from .LL1Wrapper import wrap_ll1
-from .SpecGrammar import create_spec_grammar
+from .create_spec_grammar import create_spec_grammar
 from plcc.load_spec.parse_spec.parse_syntactic_spec.parse_syntactic_spec import parse_syntactic_spec
 from .errors import InvalidSymbolException, InvalidSyntacticSpecException
