+++ src/plcc/load_spec/validate_spec/validate_syntactic_spec/validate_ll1/generate_first_sets_test.py
@@ -1,8 +1,8 @@
 from pytest import raises
 from plcc.load_spec.load_rough_spec.parse_dividers import parse_dividers
 from plcc.load_spec.load_rough_spec.parse_lines import Line
-from .FirstSets import generate_first_sets
-from .SpecGrammar import create_spec_grammar
+from .generate_first_sets import generate_first_sets
+from .create_spec_grammar import create_spec_grammar
 from plcc.load_spec.parse_spec.parse_syntactic_spec.parse_syntactic_spec import parse_syntactic_spec
 from .errors import LeftRecursionException
